Replaced animation event for slow with code
Removed the animation and animators associated with increasing linear damping of the virus with code that implements this functionality via coroutines. This is to make way for the new animation event which will occur with the hologram. Additionally, adjusted the wall count system such that in the event that a user spawns in already "within" (in regards to their collision box) a wall, their wall count will not then become negative. If this would happen at any point, it is instead set to 0.